+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 **added**
 - performance timings to physics and rendering
 - tried to add vulkan timestamps 
+- camera struct with projection and view matrices
 
 **changed**
 - disable validation layers on release
@@ -17,9 +18,6 @@
 - fixed vulkan render pass dependency flag validation error
 - temporary fix for msvc compilers, disabled logging since there was a compilation error
 - unsupported validations layers no longer crash the program, just issue a warning
-
-**up next**
-- projection and view matrices
 
 #### 0.3.4 - all about that state c: - _18 12 2021_